requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
SQLAlchemy==2.0.20
Werkzeug==2.3.7
Jinja2==3.1.2
pandas==2.1.0
numpy==1.25.2
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
pdfkit==1.0.0
weasyprint==60.1
Pillow==10.0.0
markdown==3.4.4
requests==2.31.0
httpx==0.24.1
